Setup Github Actions
This configures github actions to test `mkchlog` as well as run it on
its own repository. Also moved `.mkchlog.yml`, which was used in test,
to `tests/mkchlog.yml` and created custom `.mkchlog.yml` that's used in
this project.

The new `.mkchlog.yml` is heavily inspired by the original example with
more sections, so we're more flexible in the future. Includes a section
used in this commit. :)

changelog:
	section: dev
	title-is-enough: true