Support building on Debian Bookworm
This change lowers the requirements for dependencies so that the code
compiles on Rust 1.63 which is in Debian Bookworm. Further, the
dependencies are lowered such that the packages vendored in Debian
Bookworm can be used directly.

This uses version ranges so that the newest crates can still be used
(they didn't break our code).

changelog:
	section: features
	title-is-enough: true